Do testing with a multibroker Kafka cluster
Sets up the tests with a multi-broker Kafka cluster. This setup
unfortunately required 3 explicit brokers in the docker-compose.yml, but
due to the current mechanics of docker-compose, this is necessary to
have all 3 Kafka brokers be accessible to the host machine, since they
all require being mapped to different ports.